MNIST-Bags_Experiments/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
blas=1.0=mkl
ca-certificates=2020.6.20=hecda079_0
certifi=2020.6.20=py38h32f6830_0
cudatoolkit=10.2.89=hfd86e86_1
freetype=2.10.2=he06d7ca_0
joblib=0.16.0=pypi_0
jpeg=9d=h516909a_0
lcms2=2.11=hbd6801e_0
ld_impl_linux-64=2.35=h769bd43_9
libffi=3.2.1=he1b5a44_1007
libgcc-ng=9.3.0=h24d8f2e_16
libpng=1.6.37=hed695b0_2
libstdcxx-ng=9.3.0=hdf63c60_16
libtiff=4.1.0=hc7e4089_6
libwebp-base=1.1.0=h516909a_3
llvm-openmp=10.0.1=hc9558a2_0
lz4-c=1.9.2=he1b5a44_3
mkl=2020.2=256
mkl-service=2.3.0=py38h1e0a361_1
mkl_fft=1.2.0=py38hab2c0dc_1
mkl_random=1.1.1=py38hcb8c335_0
ncurses=6.2=he1b5a44_1
ninja=1.10.1=hc9558a2_1
numpy=1.19.1=py38hbc911f0_0
numpy-base=1.19.1=py38hfa32c7d_0
olefile=0.46=py_0
openssl=1.1.1g=h516909a_1
pillow=7.2.0=py38h9776b28_1
pip=20.2.3=py_0
python=3.8.5=h1103e12_8_cpython
python_abi=3.8=1_cp38
pytorch=1.6.0=py3.8_cuda10.2.89_cudnn7.6.5_0
readline=8.0=he28a2e2_2
scikit-learn=0.23.2=pypi_0
scipy=1.5.2=pypi_0
setuptools=49.6.0=py38h32f6830_1
six=1.15.0=pyh9f0ad1d_0
sklearn=0.0=pypi_0
sqlite=3.33.0=h4cf870e_0
threadpoolctl=2.1.0=pypi_0
tk=8.6.10=hed695b0_0
torchvision=0.7.0=py38_cu102
tqdm=4.49.0=pypi_0
wheel=0.35.1=pyh9f0ad1d_0
xz=5.2.5=h516909a_1
zlib=1.2.11=h516909a_1009
zstd=1.4.5=h6597ccf_2
